requirements.txt
Flask==0.12.2
Flask-Assets==0.12
Flask-SQLAlchemy==2.3.2
SQLAlchemy==1.1.15

requests==2.18.4

# Postgres support
psycopg2
python-dateutil
